Refactored Driver Management and added Driver Management's list Function
