+++ snackpacks-server/API - Driver Management/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "snackpacks-server",
+  "version": "1.0.0",
+  "description": "the server package",
+  "main": "index.js",
+  "directories": {
+    "test": "tests"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "mysql": "^2.16.0"
+  }
+}
