--- snackpacks-server/API - Driver Management/src/orderConnect.js
@@ -1,74 +1,0 @@
-//orderConnect.js
-//NODEJS ES5
-//Purpose: abstract the process of working with the database
-
-//Required libraries
-//Custom libs
-var Order = require("./Order");
-
-//Other
-var mysql = require('mysql');
-
-class orderConnector{
-	//orderConnector constructor
-	constructor(){
-		this.host = "snackpacksdb.cawigtgndeba.us-east-2.rds.amazonaws.com";
-		this.user = "snackpacks";
-		this.password = "e7p$yYzRa&RR_46u";
-		this.port = 3306;
-	}
-
-	//Orders
-
-	getOrders(){
-		return new Promise((resolve, reject) => {
-			var connection = mysql.createConnection({host:this.host, user:this.user, password:this.password, port:this.port});
-			//Start the descent into callback hell
-			connection.connect(function(err) {
-				if (err) reject(err);
-				//callback to send query
-				//Instead of trying to iterate thru an array
-				connection.query(`SELECT * FROM snackpacks.orders`, function(err, result, fields){
-					if (err) reject(err);
-					//callback to end connection
-					connection.end(function(err) {
-						if (err) reject(err);
-
-						//Iterate through JSON object returned by SQL query and add new SnackPack objects to list_snackpacks
-						var orderList=[];
-						for(var r in result){
-							var orderItem = result[r];
-							orderList.push(new Order(orderItem.id, orderItem.name, orderItem.price, orderItem.calories, orderItem.allergens));
-						}
-						resolve(orderList);
-					});
-				});
-			});
-		});
-	}
-
-	getOrderByID(id){
-		return new Promise((resolve, reject) => {
-			var connection = mysql.createConnection({host:this.host, user:this.user, password:this.password, port:this.port});
-			//Start the descent into callback hell
-			connection.connect(function(err) {
-				if (err) reject(err);
-				//callback to send query
-				//Instead of trying to iterate thru an array
-				connection.query(`SELECT * FROM snackpacks.Orders where id=${id}`, function(err, result, fields){
-					if (err) reject(err);
-					//callback to end connection
-					connection.end(function(err) {
-						if (err) reject(err);
-						//Iterate through JSON object returned by SQL query and add new SnackPack objects to list_snackpacks
-						var orderItem = result[0];
-						resolve(orderItem);
-					});
-				});
-			});
-		});
-	}
-}
-
-//Allows module to be exposed
-module.exports = orderConnector;
